pass certain query parameters to arethusa
Pass the query parameter `w` to the Arethusa initializer.
For example, `/histories-40-59/1?w=1` will pass `w: 1` to Arethusa which
will highlight the first word.

When extra arguments are passed to Arethusa, the buttons in the
`ControlPanel` are regular `<a>` links instead of React Router
`<Link>`s. This makes the browser's back button work correctly with the
query parameters.