add embed view
For every treebank, add an "embed" view that can be accessed by visiting
`/embed/:publication/:chunk`

For example, `/on-the-murder-of-eratosthenes-1-50/1` in the embed view
would be `/embed/on-the-murder-of-eratosthenes-1-50/1`.

This view doesn't have any of the metadata; it just shows the treebank.
There is a link at the bottom that links to the full page.