sidepanel morphology layout
When a treebank is viewed with `?config=sidepanel` then show the
sidepanel with morphology tab.

Unlike with the `?w=` parameter, this `?config=` parameter does not
cause the control panel buttons to become real links. This also means
that the `?config=` argument is carried over when `First`, `Next`, etc.
is pressed.