move openssl call to before_deploy in travis
The Travis build fails if somebody clones the repository and makes a
pull request. It fails at the point it decrypts `.travis-deploy-key`
because the secrets are not provided. (This is the correct
behavior. If it were otherwise somebody could make a pull request with
an updated `.travis.yml` that exposes all the secret information.)

Since the `.travis-deploy-key` is only used for deployment on `master`
there's no reason for this to break the build. Moving the decryption
from `before_install` to `before_deploy` should fix the issue.